# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Swagger API",
        "description": "Laravel Swagger API Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server - Development"
        }
    ],
    "paths": {
...
        "/api/task": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Obtener la lista de Task",
                "description": "Obtiene una lista de todos las tareas disponibles",
                "operationId": "getTask",
                "responses": {
                    "200": {
                        "description": "Lista obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista obtenida correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al obtener la lista: Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Crear un nuevo Task",
                "description": "Crea una nueva tarea con los datos proporcionados",
                "operationId": "createTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "title",
                                    "status",
                                    "priority",
                                    "due_date"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Tarea de ejemplo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Esto es opcional",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "medium"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "example": "2025-06-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dato creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dato creado correctamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Los datos proporcionados no son válidos."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al crear: Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}